Add dedicated chunking script
